backend/requirements.txt
boto3
pydantic
fastapi
uvicorn
python-dotenv
